+++ saleor/graphql/discount/descriptions.py
@@ -1,0 +1,7 @@
+class VoucherDescriptions:
+    TYPE = (
+        "Voucher type: PRODUCT, CATEGORY SHIPPING or ENTIRE_ORDER. "
+        "Deprecated fields: "
+        "PRODUCT, COLLECTION, CATEGORY use SPECIFIC_PRODUCT instead. "
+        "VALUE use ENTIRE_ORDER instead."
+    )