+++ saleor/static/dashboard-next/discounts/translations.ts
@@ -1,13 +1,13 @@
 import i18n from "../i18n";
-import { VoucherType } from "../types/globalTypes";
+import { VoucherTypeEnum } from "../types/globalTypes";
 
 export const translateVoucherTypes = () => ({
-  [VoucherType.CATEGORY]: i18n.t("Selected Categories"),
-  [VoucherType.COLLECTION]: i18n.t("Selected Collections"),
-  [VoucherType.PRODUCT]: i18n.t("Selected Products"),
-  [VoucherType.SHIPPING]: i18n.t("Shipment"),
-  [VoucherType.ENTIRE_ORDER]: i18n.t("Entire order"),
-  [VoucherType.SPECIFIC_PRODUCT]: i18n.t(
+  [VoucherTypeEnum.CATEGORY]: i18n.t("Selected Categories"),
+  [VoucherTypeEnum.COLLECTION]: i18n.t("Selected Collections"),
+  [VoucherTypeEnum.PRODUCT]: i18n.t("Selected Products"),
+  [VoucherTypeEnum.SHIPPING]: i18n.t("Shipment"),
+  [VoucherTypeEnum.ENTIRE_ORDER]: i18n.t("Entire order"),
+  [VoucherTypeEnum.SPECIFIC_PRODUCT]: i18n.t(
     "Selected Products, Collections and Categories"
   )
 });