+++ saleor/static/dashboard-next/discounts/types/VoucherCataloguesAdd.ts
@@ -2,7 +2,7 @@
 /* eslint-disable */
 // This file was automatically generated and should not be edited.
 
-import { CatalogueInput, VoucherDiscountValueType, VoucherType } from "./../../types/globalTypes";
+import { CatalogueInput, VoucherDiscountValueType, VoucherTypeEnum } from "./../../types/globalTypes";
 
 // ====================================================
 // GraphQL mutation operation: VoucherCataloguesAdd
@@ -141,7 +141,7 @@ export interface VoucherCataloguesAdd_voucherCataloguesAdd_voucher {
   discountValue: number;
   countries: (VoucherCataloguesAdd_voucherCataloguesAdd_voucher_countries | null)[] | null;
   minAmountSpent: VoucherCataloguesAdd_voucherCataloguesAdd_voucher_minAmountSpent | null;
-  type: VoucherType;
+  type: VoucherTypeEnum | null;
   code: string;
   used: number;
   applyOncePerOrder: boolean;