+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Use "esModuleInterop" flag in tsconfig to simplify imports - #4372 by @dominik-zeglen
 - Use hooks instead of class component in forms - #4374 by @dominik-zeglen
 - Drop csrf token header from API client - #4357 by @dominik-zeglen
+- Fix duplicate voucher enum in schema, `VoucherType` enum type was dropped, use `VoucherTypeEnum` instead - #4385 by @NyanKiyoshi
 
 ## 2.7.0